+++ src/Preview.tsx
@@ -1,7 +1,7 @@
 import { CertificatesPreview } from "./components/Certificates";
 import { EducationPreview } from "./components/education/education-preview";
 import { EmploymentPreview } from "./components/Employment";
-import { InterestsPreview } from "./components/Interests";
+import { InterestsPreview } from "./components/interests/interests";
 import { PersonalDetailsPreview } from "./components/personal-details/personal-details-preview";
 import { QualificationsPreview } from "./components/qualifications/qualifications-preview";
 import { ReferencesPreview } from "./components/References";
